Don't colorize groups if given --nocolor
ack was colorizing captured groups even if --nocolor was given.  Fixed
this by moving up the "if ( $color )" guard to cover all the
colorizing bits in print_line_with_options.  I also added a test to
t/ack-interactive.t to cover this bug.

ack-color.t had a test that was only passing because of this bug where
captured groups would always be colored; that test was fixed by just
adding --color to its args.