Changes to config finder to install on Windows.
The tests were failing on Windows if certain global files didn't exist.
Now we check for them and create them if they don't, then delete them
when we're done with the test. This *should* mean that ack will install
properly on Windows now (!), but it would be nice if some more people
could verify this.