Add --color option to timings script
Setting the --color option causes the timings script to
color the timings based on whether they were faster or slower than
the previous ack tested.  Right now it's a simple faster/slower
distinction, but in the future we could take advantage of larger
colorsets that more modern terminals provide if we want.