Add fork resources impl
This implementation iterates over directory structure in a subprocess,
yielding entries via a pipe.  The idea is that traversal and filtration
can happen while the master process is searching another file.

This implementation didn't do terribly well in benchmarks (it didn't
result in any real improvement in speed), but that was on a VM using
a single CPU.  I'll try on real hardware later.