Fix squash and build process
This addresses issue #23.  It solves the problem by putting ack
at the end of the generated file rather than the beginning.  The
reason I did this is although it was a low priority, it caused
confusing behavior when using module-level variables, and it
could cause untold confusion in prospective contributors.