Slurp files into a huge chunk to search for no frills searches
In most cases, a user is going to do something like this:

  ack foo $DIR

Where many files in $DIR don't have foo in them at all, and those
that do will probably have them in a lower percentage of their
contents.  So we try to spend more time in the regex engine rather
than Perl land.