Reduce unnessecary big regexps
Since the regular expression should contain [pt]+, means that there
might be an arbitrary number of t:s at the end. The following t+ means
the regexp should end with a t. Removing the + means the t:s prior is
just contained in the [pt]+

This reduces the file size with a couple of bytes, which, even though
it's really unlikely, means having space for this app or not.

THPPPTPTPTPTPTT!